console: Separate routing configuration to include gateway visibility
