+++ pkg/webui/console/views/admin-packet-broker/network-routing-policies.js
@@ -19,12 +19,12 @@ import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
 
 import PacketBrokerNetworksTable from '@console/containers/packet-broker-networks-table'
 
-import m from './messages'
+import sharedMessages from '@ttn-lw/lib/shared-messages'
 
 const NetworkRoutingPoliciesView = () => {
   useBreadcrumbs(
     'admin-panel.packet-broker.networks',
-    <Breadcrumb path={'/admin-panel/packet-broker/networks'} content={m.networks} />,
+    <Breadcrumb path={'/admin-panel/packet-broker/networks'} content={sharedMessages.networks} />,
   )
 
   return <PacketBrokerNetworksTable />
