+++ pkg/webui/locales/en.json
@@ -252,6 +252,7 @@
   "console.components.gateway-api-keys-modal.index.description": "Note: After closing this window, these API keys will not be accessible for download anymore. Please make sure to download and store them now.",
   "console.components.gateway-api-keys-modal.index.downloadLns": "Download LNS key",
   "console.components.gateway-api-keys-modal.index.downloadCups": "Download CUPS key",
+  "console.components.gateway-visibility-form.index.saveDefaultGatewayVisibility": "Save default gateway visibility",
   "console.components.location-form.index.deleteAllLocations": "Delete all location data",
   "console.components.location-form.index.deleteFailure": "An error occurred and the location could not be deleted",
   "console.components.location-form.index.deleteLocation": "Remove location data",
@@ -354,6 +355,7 @@
   "console.components.routing-policy-form.index.saveDefaultPolicy": "Save default policy",
   "console.components.routing-policy-form.index.useDefaultPolicy": "Use default routing policy for this network",
   "console.components.routing-policy-form.index.useSpecificPolicy": "Use network specific routing policy",
+  "console.components.routing-policy-form.index.doNotUseADefaultPolicy": "Do not use a default routing policy for this network",
   "console.components.routing-policy-form.index.doNotUseAPolicy": "Do not use a routing policy for this network",
   "console.components.uplink-form.uplink-form.simulateUplink": "Simulate uplink",
   "console.components.uplink-form.uplink-form.payloadDescription": "The desired payload bytes of the uplink message",
@@ -683,6 +685,7 @@
   "console.views.admin-packet-broker.messages.packetBrokerDisabledDesc": "The Things Stack is not set up to use Packet Broker. Please refer to the documentation link above for instructions on how to set up The Things Stack for peering with Packet Broker.",
   "console.views.admin-packet-broker.messages.enablePacketBroker": "Enable Packet Broker",
   "console.views.admin-packet-broker.messages.packetBrokerRegistrationDesc": "Enabling will allow other networks to send traffic to you as well as you forwarding traffic to them, based on the exact routing policy.",
+  "console.views.admin-packet-broker.messages.routingConfig": "Routing configuration",
   "console.views.admin-packet-broker.messages.network": "Network: {network}",
   "console.views.admin-packet-broker.messages.listNetwork": "List my network in Packet Broker publicly",
   "console.views.admin-packet-broker.messages.listNetworkDesc": "Public listing will make it easier for other network operators to set up routing policies for your network. Hence public listing is generally recommended.",
@@ -1384,7 +1387,7 @@
   "lib.shared-messages.sessions": "Sessions",
   "lib.shared-messages.setLoRaCloudToken": "Set LoRa Cloud token",
   "lib.shared-messages.settings": "Settings",
-  "lib.shared-messages.setupSubTitle": "Setup",
+  "lib.shared-messages.setup": "Setup",
   "lib.shared-messages.shareGatewayInfo": "Share gateway information",
   "lib.shared-messages.skipCryptoDescription": "Skip decryption of uplink payloads and encryption of downlink payloads",
   "lib.shared-messages.skipCryptoPlaceholder": "Encryption/decryption disabled",
