+++ pkg/webui/locales/ja.json
@@ -252,6 +252,7 @@
   "console.components.gateway-api-keys-modal.index.description": "注意：このウィンドウを閉じると、これらのAPIキーはダウンロードできなくなります。必ずダウンロードして保存してください！",
   "console.components.gateway-api-keys-modal.index.downloadLns": "LNSキーをダウンロード",
   "console.components.gateway-api-keys-modal.index.downloadCups": "CUPSキーをダウンロード",
+  "console.components.gateway-visibility-form.index.saveDefaultGatewayVisibility": "",
   "console.components.location-form.index.deleteAllLocations": "すべての位置情報を削除",
   "console.components.location-form.index.deleteFailure": "エラーが発生し、その場所を削除できませんでした",
   "console.components.location-form.index.deleteLocation": "場所入力を削除",
@@ -354,6 +355,7 @@
   "console.components.routing-policy-form.index.saveDefaultPolicy": "",
   "console.components.routing-policy-form.index.useDefaultPolicy": "このネットワークでデフォルトのルーティングポリシーを使用",
   "console.components.routing-policy-form.index.useSpecificPolicy": "ネットワーク固有のルーティングポリシーを使用",
+  "console.components.routing-policy-form.index.doNotUseADefaultPolicy": "",
   "console.components.routing-policy-form.index.doNotUseAPolicy": "このネットワークには、ルーティングポリシーを使用しないでください",
   "console.components.uplink-form.uplink-form.simulateUplink": "アップリンクのシミュレーション",
   "console.components.uplink-form.uplink-form.payloadDescription": "アップリンクメッセージの希望するペイロードバイト数",
@@ -683,6 +685,7 @@
   "console.views.admin-packet-broker.messages.packetBrokerDisabledDesc": "",
   "console.views.admin-packet-broker.messages.enablePacketBroker": "",
   "console.views.admin-packet-broker.messages.packetBrokerRegistrationDesc": "ご自宅のネットワークから、またはご自宅のネットワークへのピアリングを可能にするには、ご自宅のネットワークを登録する必要があります。これにより、お客様のネットワークがPacket Brokerに認識され、お客様のネットワークのピアリング動作を設定することができます",
+  "console.views.admin-packet-broker.messages.routingConfig": "",
   "console.views.admin-packet-broker.messages.network": "ネットワーク: {network}",
   "console.views.admin-packet-broker.messages.listNetwork": "ネットワーク公開で一覧表示",
   "console.views.admin-packet-broker.messages.listNetworkDesc": "ネットワークをリストアップすることで、他のネットワーク管理者があなたのネットワークを見ることができます。これにより、他のネットワーク管理者は、あなたのネットワークでルーティングポリシーを簡単に設定することができます",
@@ -1384,7 +1387,7 @@
   "lib.shared-messages.sessions": "",
   "lib.shared-messages.setLoRaCloudToken": "",
   "lib.shared-messages.settings": "",
-  "lib.shared-messages.setupSubTitle": "",
+  "lib.shared-messages.setup": "",
   "lib.shared-messages.shareGatewayInfo": "",
   "lib.shared-messages.skipCryptoDescription": "",
   "lib.shared-messages.skipCryptoPlaceholder": "",
